Add a vlModal directive and service to control vlModals
See #77, #82

This PR adds a vl-modal directive, and creates a VlModals service through
which an individual vl-modal can be opened or closed by ID. This removes
some of the scope variables needed by the existing modal solution (for
the bookmarks modal), and encapsulates the modal's visibility inside
the vl-modal's isolate scope.

Making this change to the bookmarks sets us up to use it for the dataset
selection modal: because the bookmarks scope situation is slightly more
straightforward, it makes for a good testbed.

Still TODO before this merges are the unit tests, which I have broken
due to removing the visibility logic from the bookmarks directive.

This PR accompanies vega/voyager#192 and the two should be merged in
tandem in order for voyager's bookmarks to continue working.