Abstract modal close button into a directive
Moving the close button into a directive frees the link function of a
modal from itself depending on or interacting with the parent
directive's modalController: they can just include a modalCloseButton
somewhere in their markup, and the modalCloseButton will handle all
the interaction with the parent directive.

The modal close button also provides an interface by which a directive
can register a handler that will run when that button is used to close
the modal.