Organize the dataset add modal with a tab directive
![modal](https://cloud.githubusercontent.com/assets/442115/10324659/554d1e42-6c59-11e5-8cb3-9da6ee64f4f1.gif)

This implements the tab directive approach @vlandham originally proposed
to organize and simplify the layout of the modal directive. I spent a
bit of time experimenting with the tabs on the side, as we had it in
the original wireframe, and decided the tabs on the top were more idiomatic
and readily accessible.