Only use customized JSON3-compactstringify within compactJSON filter
This library was globally overwriting the JSON object on window,
which significantly decreased the performance of the application
in some browsers because the try/catch within the custom code
could not be optimized.

Test case: load a ~300kb dataset with 1000 rows, and profile the
performance of adding the dataset with customized JSON3 and with
the native JSON object.

Before this PR:

![image](https://cloud.githubusercontent.com/assets/442115/10462479/f1e8e4ea-71ac-11e5-8f51-62ac32953d42.png)

Initial dataset parsing time 3.03 seconds

After this PR:

![image](https://cloud.githubusercontent.com/assets/442115/10462318/12d251e2-71ac-11e5-8fb3-fba7a8b84e36.png)

Initial dataset parsing time 1.13 seconds