Defer tool popup UI to signficantly improve first-render
Steps to reproduce the original issue:

1. Drag-and-drop or paste-load a medium-size dataset (~100 rows)
2. Click "Add Dataset"
3. Mouse over a left-column UI element to trigger chart re-rendering
4. Observe that chart rendering is extremely sluggish

Background:

Currently, the pop-up debug UI for vega lite plots has to be rendered
at the same time as the rest of the debug UI controls, and cannot be
deferred until needed with an `ng-if`, because the Drop requires the
content element to be present.

![image](https://cloud.githubusercontent.com/assets/442115/10466728/cfd94a86-71c3-11e5-95c8-99decc1cf978.png)

While this may not seem like a huge issue, because the debug popup uses
the `compactJSON` filter, this front-loaded rendering of debug UI that
may never be used is a *significant* bottleneck on render.

Status quo:

![image](https://cloud.githubusercontent.com/assets/442115/10466418/3d441558-71c2-11e5-91a5-0da64bc88456.png)

Initial render takes **8.89 seconds**, much of which is tied up in
compactjson's inefficient serialization and deserialization.

After lazy-initialization of Drop popup contents:

![image](https://cloud.githubusercontent.com/assets/442115/10466085/4b30aa34-71c0-11e5-84d9-4fa07a43f174.png)

Initial render takes **1.33 seconds**

Removing the processing burden of rendering this data, and iteratively
computing the serialized versions of the datasets, will allow the initial
render to complete in up to **14%** of the time it currently takes.

These savings are accomplished by moving the rendering of the Drop,
and the rendering of that drop's contents, into its own directive.
That directive's visibility is controlled via `ng-if`, so it will not
be evaluated until the user mouses over the debug UI for a plot.

This contextual deferred rendering allows us to completely skip all of
the debug UI rendering during the initial render pass, which gets all
the graphs on the page in much less time than they currently take.