Permit <modal> directive to auto-open based on a directive attribute
In order to support the welcome modal, two approaches could be used:
either the controller for the template in which the modal occurs could
explicitly `Modals.open('welcome-modal')`, or else the welcome modal
could be augmented with an attribute that determines whether the modal
will render as open by default instead of closed.

This PR supports the latter interaction as it encapsulates the display-
oriented state (whether to show or not) into the directive, rather
than depending on a service for display state.