Switch myria dataset selector to an ordinary dropdown
Fixes #101

After consulting with colleagues, we have no good solution for
forcing a custom-rendered select UI (as is generated by ui-select)
to overflow the edges of a div with overflow: scroll or hidden
selected. As such, the only viable solutions to this problem would
be to convert it to a list in the DOM, as is done on the change data
tab; or to revert to a browser-default select box.

In the interest of time, I opted for the latter.

This should be reviewed, I'm particularly uncertain about what was
happening with the `$select.search` value previously; but as is
demonstrated in the gif below, this does permit the user to switch
Myria datasets without the list being cut off.

![select](https://cloud.githubusercontent.com/assets/442115/10554114/8a96e2ea-742e-11e5-9f29-d0ba3743d662.gif)